functions/requirements.txt
functions-framework==3.*
firebase-admin==6.*
google-cloud-firestore==2.14.0
google-events